Added same access control logic to the deleteForm. Right now a user can only delete forms belongs to him/her (Only two genders?????).

+++ src/main/java/com/aozbek/form/service/FormService.java
@@ -48,6 +48,10 @@ public class FormService {
     public void deleteForm(String formId) {
         Form form = formRepository.getFormById(formId)
                 .orElseThrow(() -> new NoSuchElementException("No form exist with this id."));
+        String userId = authService.getCurrentUser().getId();
+        if (!(form.getUserId().equals(userId))) {
+            throw new UnauthorizedAccessException();
+        }
         formRepository.delete(form);
     }
 }
